+++ src/main/java/com/guidewire/gigsuraksha/service/incomeprofileservice.java
@@ -1,11 +1,12 @@
 package com.guidewire.gigsuraksha.service;
+import java.math.BigDecimal;
 import java.time.LocalDate;
 import java.util.UUID;
 public interface incomeprofileservice {
 
 	void buildProfile(UUID partnerId);
 
-    Double predictDailyIncome(UUID partnerId, LocalDate date);
+    BigDecimal predictDailyIncome(UUID partnerId, LocalDate date);
 
     void refreshModel(UUID partnerId);
 	}
