+++ src/main/java/com/guidewire/gigsuraksha/service/trustscoreservice.java
@@ -1,5 +1,6 @@
 package com.guidewire.gigsuraksha.service;
 
+import java.math.BigDecimal;
 import java.util.UUID;
 
 
@@ -7,7 +8,7 @@ public interface trustscoreservice {
 
 	void applyDelta(UUID partnerId, String reason, Double amount);
 
-    Double getCurrentScore(UUID partnerId);
+    BigDecimal getCurrentScore(UUID partnerId);
 
     void checkSuspension(UUID partnerId);
